test(popover): migrate size
